+++ docs/docs/documenting-others.tsx
@@ -1,29 +1,16 @@
 import * as React from 'react';
-import { useQueryClient } from 'react-query';
-import {
-  BrowserWindow,
-  MarkdownArticle,
-  mdxComponents,
-} from 'react-showroom/client';
+import { mdxComponents } from 'react-showroom/client';
+import { MarkdownDocRoute } from '../components/markdown-doc-route';
 import * as markdownExampleResult from './markdown-example.mdx';
 import markdownExampleSource from './markdown-example.mdx?showroomRaw';
 import markdownExampleCodeblocks from './markdown-example.mdx?showroomRemarkCodeblocks';
+import * as markdownExampleImports from './markdown-example.mdx?showroomRemarkDocImports';
+import { useSetCompilationCaches } from './set-compilation-caches';
 
 const { pre: Pre, code: Code } = mdxComponents;
 
 export const MarkdownExampleSource = () => {
-  const queryClient = useQueryClient();
-
-  React.useEffect(() => {
-    [markdownExampleCodeblocks].forEach((blocks) => {
-      Object.keys(blocks).forEach((soureCode) =>
-        queryClient.setQueryData(
-          ['codeCompilation', soureCode],
-          blocks[soureCode]
-        )
-      );
-    });
-  }, []);
+  useSetCompilationCaches([markdownExampleCodeblocks]);
 
   return (
     <Pre>
@@ -35,18 +22,16 @@ export const MarkdownExampleSource = () => {
 };
 
 export const MarkdownExampleResult = () => (
-  <BrowserWindow url="http://localhost:6969/content" className="mb-4">
-    <section className="px-6">
-      <MarkdownArticle
-        section={{
-          type: 'markdown',
-          Component: markdownExampleResult.default,
-          title: markdownExampleResult.title,
-          slug: 'content',
-          headings: markdownExampleResult.headings,
-          frontmatter: {},
-        }}
-      />
-    </section>
-  </BrowserWindow>
+  <MarkdownDocRoute
+    data={{
+      type: 'markdown',
+      Component: markdownExampleResult.default,
+      title: markdownExampleResult.title,
+      slug: 'content',
+      headings: markdownExampleResult.headings,
+      frontmatter: {},
+      codeblocks: markdownExampleCodeblocks,
+      imports: markdownExampleImports.imports,
+    }}
+  />
 );