+++ packages/core/src/react.ts
@@ -5,7 +5,7 @@ import type {
   ParserOptions,
 } from 'react-docgen-typescript';
 import type { Configuration } from 'webpack';
-import { Environment } from './index';
+import { Environment, CodeBlocks } from './index';
 
 export interface ItemConfigurationWithPath {
   title?: string;
@@ -274,6 +274,7 @@ export interface ComponentDocItem {
   };
   doc: null | ComponentType<any>;
   imports: Record<string, any>;
+  codeblocks: CodeBlocks;
 }
 
 export interface ReactShowroomComponentSection {
@@ -303,6 +304,8 @@ export interface ReactShowroomMarkdownSection {
   slug: string;
   frontmatter: ReactShowroomMarkdownFrontmatter;
   headings: Array<ReactShowroomMarkdownHeading>;
+  imports: Record<string, any>;
+  codeblocks: CodeBlocks;
 }
 
 interface ReactShowroomLinkSection {