+++ packages/react-showroom/client/app.tsx
@@ -1,14 +1,15 @@
 import { IdProvider } from '@radix-ui/react-id';
+import { css, QueryParamProvider } from '@showroomjs/ui';
+import cx from 'classnames';
 import * as React from 'react';
 import { Route, Switch, useLocation } from 'react-router-dom';
 import sections from 'react-showroom-sections';
 import Wrapper from 'react-showroom-wrapper';
-import { ComponentDocArticle } from './components/component-doc-article';
-import { DetailsPageContainer } from './components/details-page-container';
-import { MarkdownArticle } from './components/markdown-article';
 import { CodeThemeContext } from './lib/code-theme-context';
-import { QueryParamProvider } from './lib/use-query-params';
+import { SubRootRoute } from './lib/routing';
+import { ComponentDocRoute } from './pages/component-doc-route';
 import { DefaultHomePage } from './pages/index';
+import { MarkdownRoute } from './pages/markdown-route';
 import { colorTheme, THEME } from './theme';
 
 export const App = () => {
@@ -42,7 +43,7 @@ export const App = () => {
   return (
     <Wrapper>
       <IdProvider>
-        <div className={colorTheme}>
+        <div className={cx(colorTheme, wrapper())}>
           <QueryParamProvider>
             <CodeThemeContext.Provider value={THEME.codeTheme}>
               <Switch>
@@ -65,38 +66,24 @@ export const App = () => {
 
                   if (section.type === 'component') {
                     return (
-                      <Route path={`/${section.slug}`} key={section.slug}>
-                        <DetailsPageContainer
-                          title={section.data.component.displayName}
-                          description={section.data.component.description}
-                        >
-                          <ComponentDocArticle doc={section} />
-                        </DetailsPageContainer>
-                      </Route>
+                      <SubRootRoute
+                        path={`/${section.slug}`}
+                        key={section.slug}
+                      >
+                        <ComponentDocRoute section={section} />
+                      </SubRootRoute>
                     );
                   }
 
                   if (section.type === 'markdown') {
                     return (
-                      <Route
+                      <SubRootRoute
                         path={`/${section.slug}`}
                         exact={section.slug === ''}
                         key={section.slug}
                       >
-                        <DetailsPageContainer
-                          title={
-                            section.slug === '' ? undefined : section.title
-                          }
-                          description={section.frontmatter.description}
-                          hideSidebar={section.frontmatter.hideSidebar}
-                          hideHeader={section.frontmatter.hideHeader}
-                        >
-                          <MarkdownArticle
-                            section={section}
-                            center={!section.frontmatter.hideSidebar}
-                          />
-                        </DetailsPageContainer>
-                      </Route>
+                        <MarkdownRoute section={section} />
+                      </SubRootRoute>
                     );
                   }
 
@@ -113,3 +100,7 @@ export const App = () => {
     </Wrapper>
   );
 };
+
+const wrapper = css({
+  height: '100%',
+});