+++ packages/react-showroom/client/exposes.tsx
@@ -1,8 +1,15 @@
+export { getCompilationKey } from '@showroomjs/core';
+export type { CodeBlocks } from '@showroomjs/core';
+export { QueryParamProvider } from '@showroomjs/ui';
 export { BrowserWindow } from './components/browser-window';
 export type { BrowserWindowProps } from './components/browser-window';
+export { ComponentDataProvider } from './components/component-data-provider';
 export { ComponentDocArticle } from './components/component-doc-article';
 export { ComponentMeta } from './components/component-meta';
 export type { ComponentMetaProps } from './components/component-meta';
 export { MarkdownArticle } from './components/markdown-article';
+export { MarkdownDataProvider } from './components/markdown-data-provider';
 export { mdxComponents } from './components/mdx-components';
 export { Seo as Head } from './components/seo';
+export { SubRootRoute } from './lib/routing';
+export { StandaloneEditor } from './pages/standalone-editor';