+++ packages/ui/src/lib/create-named-context.ts
@@ -1,0 +1,10 @@
+import { createContext } from 'react';
+
+export const createNameContext = <ContextType>(
+  name: string,
+  defaultValue: ContextType
+) => {
+  const context = createContext(defaultValue);
+  context.displayName = name;
+  return context;
+};